Started work on multiplayer
Controls are mixed up i need to breakdown the movement script into a few different scripts

+++ Assets/FishNet/Runtime/Managing/Scened/LoadUnloadDatas/SceneLoadData.cs
@@ -56,7 +56,7 @@ namespace FishNet.Managing.Scened
         /// </summary>
         /// <param name = "sceneHandle">Scene to load by handle.</param>
         /// <param name = "sceneName">Scene to load by name.</param>
-        public SceneLoadData(int sceneHandle, string sceneName) : this(new SceneLookupData(sceneHandle, sceneName)) { }
+        public SceneLoadData(ulong sceneHandle, string sceneName) : this(new SceneLookupData(sceneHandle, sceneName)) { }
 
         /// <summary>
         /// </summary>
